Refactor: test_amber macro for code reuse
