+++ test_files/str/trim.ab
@@ -1,5 +1,5 @@
 pub fun trim(text) {
-    return "$\{{nameof text}##*( )}"
+    return unsafe $echo "{text}" | xargs$
 }
 
 main {
